refactor: Move logic for sorting arrays to prepare the rendering into `gitlab/component_documentation`
